backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.30.6
python-multipart==0.0.12
pydantic==2.9.2
pydantic-settings==2.6.1

sqlmodel==0.0.22
aiosqlite==0.20.0

PyPDF2==3.0.1
python-docx==1.1.2

numpy==2.1.3
scikit-learn==1.5.2